+++ Panels/WizardPanel.js
@@ -92,6 +92,20 @@ const WizardPanelBase = kind({
 		onChange: PropTypes.func,
 
 		/**
+		 * Called when a transition completes.
+		 *
+		 * @type {Function}
+		 */
+		onTransition: PropTypes.func,
+
+		/**
+		 * Called before a transition begins.
+		 *
+		 * @type {Function}
+		 */
+		onWillTransition: PropTypes.func,
+
+		/**
 		* The text for previous button.
 		*
 		* @type {String}
@@ -158,7 +172,7 @@ const WizardPanelBase = kind({
 		}
 	},
 
-	render: ({buttons, children, footer, index, total, nextButtonText, onIncrementStep, onDecrementStep, prevButtonText, reverseTransition, subtitle, title, ...rest}) => {
+	render: ({buttons, children, footer, index, total, nextButtonText, onIncrementStep, onDecrementStep, onTransition, onWillTransition, prevButtonText, reverseTransition, subtitle, title, ...rest}) => {
 		return (
 			<Panel {...rest}>
 				<Header
@@ -199,6 +213,8 @@ const WizardPanelBase = kind({
 							<ViewManager
 								arranger={SlideLeftArranger}
 								duration={400}
+								onTransition={onTransition}
+								onWillTransition={onWillTransition}
 								reverseTransition={reverseTransition}
 							>
 								{children}