+++ samples/sampler/stories/default/WizardPanel.js
@@ -20,6 +20,8 @@ storiesOf('Sandstone', module)
 			<WizardPanel
 				nextButtonText={text('nextButtonText', WizardPanel, '')}
 				onSelect={action('onSelect')}
+				onTransition={action('onTransition')}
+				onWillTransition={action('onWillTransition')}
 				prevButtonText={text('prevButtonText', WizardPanel, '')}
 			>
 				<View footer="Footer in View 1" subtitle="A subtitle for View 1" title="WizardPanel View 1">